move store folder to the src folder Relates #20
