+++ src/store/reducers/index.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from "redux";
+import { projectsReducer } from "./projects";
+import { changeReducer } from "./change";
+import { authReducer } from "./auth";
+
+export const reducer = combineReducers({
+  projects: projectsReducer,
+  change: changeReducer,
+  auth: authReducer,
+});
