+++ src/store/reducers/projects.ts
@@ -1,0 +1,13 @@
+type action = {
+  type: string;
+  payload: Array<{}>;
+};
+
+export const projectsReducer = (state: Array<{}> = [], action: action) => {
+  switch (action.type) {
+    case "STORE_DATA":
+      return action.payload;
+    default:
+      return state;
+  }
+};
